refactor: User, UserRepository 디렉토리 변경
